pyproject.toml
[project]
name = "SVAIA"
version = "0.1.0"
description = "Proyecto SVAIA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansi2html>=1.9.2",
    "argon2>=0.1.10",
    "argon2-cffi>=25.1.0",
    "bandit>=1.8.3",
    "fastapi[standard]>=0.115.12",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "flask-praetorian>=1.6.1",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "jinja2>=3.1.6",
    "logging>=0.4.9.6",
    "nvdlib>=0.8.0",
    "pkgconf>=2.4.3.post2",
    "pydantic-settings>=2.9.1",
    "pylint>=3.3.7",
    "pyright>=1.1.402",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "safety>=3.5.1",
    "semgrep>=1.125.0",
    "toml>=0.10.2",
    "uvicorn[standard]>=0.34.2",
    "websocket>=0.2.1",
]

[tool.setuptools.packages.find]
where = ["svaia"]
